reset: Allow all ledgers to reset to genesis block
This CR adds a command "peer node reset"
This command drops
 - statedb if LevelDB is being used for the state db
 - historydb if being used
 - blockstorage indexes
 - all the block files other the blockfile number zero
In addition, this truncates the blockfile number zero
to the offset where the genesis block finishes.

In other words, ledger rolls back to a state where it has
just been created as an effect of joining the channel.

Private data blockstore is not dropped because if this gets
dropped everywhere then the private data will be lost forever.

Before trucating the block file, it takes a backup of
the genesis block in a file '__backupGenesisBlockBytes' in the
same directory. It also produces another file __preResetHeight
that contains the max(blockstoreHeight,
height in __preResetHeight - if exists)

This command is intended to be fail-safe in the sense that
this should be able to be executed again if a failure or
system crash is to happen while this command is running.

Important note: if a network is using the pvtData feature
and needs to apply this command, this should be not be applied
on all the peers that share a collection, at the same time.
Further, it should be applied on the peers in select
batches such that it is appliead on the later batch peers
only after the earlier peers have reached the original block
height.

FAB-14808 #done

Change-Id: Id19677b4828c9c1fa877d4c350987aa048df73c2
Signed-off-by: manish <manish.sethi@gmail.com>
Signed-off-by: senthil <cendhu@gmail.com>
(cherry picked from commit b83ae45f610a29fae3a61849e8b526fec5b8a198)