FAB-15299 Consensus migration: Raft start sequence
Detect restart after consensus-type migration by:

1. The last block and the block before it carry a config-tx
2. In both config txs we have ConsensusType.State == MAINTENANCE
3. The ConsensusType.Type is different in these two blocks
4. The last block (orderer) metadata is empty

This part is implemented in multichannel/chainsupport.go in order
to be reusable.

In the etcdraft chain we have to change the start method.
Change the etcdraft chain.start() such that:

  isJoin = height >0 && not-migration

(instead of isJoin=height>0). "not-migration" is computed by
calling the respective method in chainsupport that detects
restart after migration. We do that in order to change the way
chain.node.start(fresh, join) is starting (vs. restarting).

Change-Id: I816ad4476955f679baaa9c0a2621399ca77a5d01
Signed-off-by: Yoav Tock <tock@il.ibm.com>