[FAB-14115] Fix Intermittent cluster test failure
The cluster deliver client has a test timeout of 1 second,
and if the timeout expires - the production code then retries
the request, which causes a test failure because the test expects
something to happen that doesn't happen if a retry takes place.

From observing the logs of some of the tests, rarely - the server
in the test, doesn't respond in a timely manner of 1 second,
and a retry takes place.

19:50:07 2019-05-08 16:48:38.212 UTC [test] tryFetchBlock ->
ERRO 0a9 Failed pulling blocks: block header is nil
19:50:07 2019-05-08 16:48:39.226 UTC [test] fetchLastBlockSeq ->
ERRO 0aa Failed receiving the latest block from 127.0.0.1:37266: didn't receive a response within 1s

I increased the timeout to 10 seconds, as a saner upper bound.

Change-Id: I68962bbed7b288ac2784d6c0bd8a0a43a5819b1c
Signed-off-by: yacovm <yacovm@il.ibm.com>