[FAB-15388] AdminOU Support
This change-set does the following:
- introduce a new MSP version, 1.4.2
- BCCSPMSP when instatiated with version 1.4.2 behaves
as follows:
-- If NodeOU support is enabled, then
the certificates in the admin folder must carry either
the client or the new admin OU. The folder
can now be empty. If the NodeOU support is not enabled
nothing changes.
-- if NodeOU support is enabled, then
when SatisfiesPrincipal evaluates an MSPPrincipal_ROLE,
it first checks that the identity carries the admin OU.
If the identity does not carry the admin OU then, the MSP
checks that it is named in the admin folder.
- Configuration load has been updated to not fail when
the admin folder is empty, the failure is triggered
later when the MSP is instantiad and in the appropriate
cases.
- What is missing is the following:
-- Enable 1.4.2 as version for the local msp
msp/mgmt/mgmt.go#loadLocalMSP must be changed.
-- Enable 1.4.2 as version for the channel msp
common/capabilities/channel.go#MSPVersion must be changed.

Change-Id: I1677c577810881ef61ffb1105ee7d6e1448b3066
Signed-off-by: Angelo De Caro <adc@zurich.ibm.com>