[FAB-15476] Reduce CouchDB exposure via reuse
Refactored lockbabsedtxmgr to use one instance of Couch.
Instead of starting a new Couch container for each test
we initialize a new ledger in the same DB. This reduces our
exposure to the number of possible timeouts on Couch
initializations to a single instance, down from > 30.

This reduced the time for testing this package by over 70%

Change-Id: If6333812872a77ea5375dcb2d9ef6ae4b49c5634
Signed-off-by: Brett Logan <Brett.T.Logan@ibm.com>