[FAB-15463] Wait for full channel membership
Tests that use the leader election process should wait for full membership
on the channel of the election before the election process starts.
Otherwise, there could be cases where there would be 2 leaders on the same
channel and the test might fail or wait longer for one of them to renounce
its leadership.

In addition, in TestLeaderElectionWithRealGossip(), a wait was added
before the leader election in the second channel.

Change-Id: I0f7dd2a66f5150bed14efbc34504fdfddae62049
Signed-off-by: ronenschafferibm <ronen.schaffer@ibm.com>