[FAB-15515] add creator footprint to couch runner
We have had to chase down CouchDB instances that get left running at the
end of unit tests suites several times now so it's just a matter of time
before we have to do it again.

In preparation for that day, let's leave a footprint behind to let us
know the code that started it. We'll do this by putting the stack trace
of the start into the environment of the container.

Change-Id: I75817d5590438f1de8240879d5aecafa7d942bc1
Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>