a new lossy Comm.Send test for the lossy behavior
TestCommSend replaces both TestBasic and TestParallelSend.

Comm.Send is both asynchronous and lossy, so the current tests which
assumed a 100% success rate were improperly designed.

Removed TestProdConstructor as it was an almost exact duplicate of TestBasic.

This doesn't fix the underlying issues in the comm code, but in
practice, the racing issues are both very unlikely, and very
recoverable if the users understand that the channels are lossy and
have no guarantees of throughput.

Refactor TestAccept with understanding of the same behavior as TestCommSend.

FAB-13539
FAB-14936
FAB-15486

Change-Id: Ia4dff1f3ee0498e481d533c4e9ef8db4c474e922
Signed-off-by: Morgan Bauer <mbauer@us.ibm.com>