Remove duplicate code snippet from tests
Preparation of ledger mgr initializer is a decently large code snippet
that gets duplicated across tests packages. Further, this may keep growing
as we pass on new dependencies. Most of the tests do not bother for almost
all the dependencies and blindly copy the code snippet.

This CR moves this code to a common place.

FAB-15771 #done

Change-Id: I03ae2aa15ab2407c5307f97b73aa6d70525ec0d6
Signed-off-by: manish <manish.sethi@gmail.com>