FAB-14724 Move ValidateMetadata to chain
Currently Consenter.MetadataValidator is stateless - it does not
require any internal state of Consenter, despite it's being a
member method. However, sometimes it requires extra info to validate
a transaction, i.e. current active nodes, which makes it "stateful".

Therefore, this CR moves this method from Consenter to Chain, so it
has access to internal state of a node.

Change-Id: I337eed748ec7169997af8175e085cbd2cf4846b9
Signed-off-by: Jay Guo <guojiannan1101@gmail.com>