FAB-14491 Collapose inproccontroller into sccp
There is no reason to maintain a separate inproc registry in addition to
the one maintained by the sysccprovider.  There is also no reason to
deploy system chaincodes over and over again, on a channel by channel
basis (they are already running, so the deployment is skipped).  This CR
kills these dead/useless codepaths.

Change-Id: I31a40baaf3e573928abee70732943660fc9fd20c
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>