FAB-14491 platforms direct to dockercontroller
Today, there is a platform registry which is created, which is handed to
the container runtime, which is handed to a platform builder, which is
invoked by the build VM req, which is passed to the dockercontroller
Build function.

This CR shorts this chain by simply initializing the dockercontroller
with a reference to a platform builder, so that the container runtime
and vm controller do not have to be involved.

Change-Id: Id39504a8937b9513695e7c677f5f209f2916c6be
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>