FAB-14491 Kill VM Provider and VM Reqs
There is currently a lot of coordination and code around passing "VM
Requests" from the chaincode container runtime, to be dispatched to the
correct VM, and then replied to back to the container runtime.  There's
really no reason these can't be direct calls, and further, there's no
reason that the VMs can't be single instance, especially as access to
them is serialized.  This CR does those things.

Change-Id: I4bc65518a677b99eaf3d575ca0f7e035186db29e
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>