FAB-11954 container start args to dockercontroller
Presently, the container runtime code generates the args, environment,
and filesystem for the container.  The container runtime should not
understand these things.  Beginning by pushing the launch args into the
container.

Note, that there's an argument to be made, that the launch args should
be abstracted into the platforms package, and that we should build an
abstraction around this to allow more easily adding platforms.  However,
our long-term goal is to kill the docker platform and use external
builders, so this additional work seems unnecessary.

Change-Id: I7cd653c22254d0b5b0f37d875d629f5c937990be
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>