FAB-11491 Pass tlsConfig instead of env
Presently, the VM interface is aligned for running arbitrary code with
arbitrary environments in a docker container.  We need this to shift to
being 'run chaincode with chaincode artifacts in an arbitrary
container'.  This CR pushes the details of translating the tlsConfig
into the docker environment to the docker controller to allow other
handling of the tls config to take place later.

Change-Id: Ifdc51b03e912523a3f4b0d13084b102908854191
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>