[FAB-16149] Check status of all proposal responses
Currently the CLI only checks the status of the
first proposal response and ignores the status of
any other proposal responses. This leads to misleading
errors about the payloads not matching, and hides
the actual error message from the failing peer(s).

This CR changes the code so that the status of all
proposal responses is checked, and results in an
error if any of them are < 200 and >= 400.

Signed-off-by: Simon Stone <sstone1@uk.ibm.com>
Change-Id: I8568b1a8ba0c9f9f9b9e81bc4c5647912933ae63