FAB-16173 avoid explicit binding to ephemeral port
On our linux based CI systems, the default ephemeral port range
(aka ip_local_port_range) is 32768-60999. Our integration test reserved
ports overlap significantly with this range and, depending on what's
going on at the time, this results in bind failures that fail the tests.

This changes the base ports to avoid conflicting with the linux defaults
in an effort to fix bind failures in integration tests.

Change-Id: I46842679984a44225c83110b616376017a425c1d
Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>