FAB-16158 Create streaming package retriever
This CR re-implements the package parsing in a streaming fashion, rather
than relying on pulling the whole code package into RAM as the previous
implementation did.  Note, the old implementation still exists, and is
still used by the non-streaming code, notably the lifecycle SCC.  We
could probably transition that code onto this interface.

Change-Id: Ib8216c57570b2f662359d4818a6614d53c4a2347
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>