FAB-16158 Fetch legacy chaincode by name:version
Presently, the legacy ccprovider fs implementation takes a name and a
version.  However, the new chaincode persistence needs a package ID.
The chaincode ID is the package ID in the new case, and the name:version
in the legacy case.  So, much of the code and shims and routing can
collapse by converting the legacy impl to operate on CCID.

Change-Id: I294bad05d67ac9f98dfcee849441e3616c156999
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>