FAB-16158 Legacy LSCC security sideeffect explicit
The current version of LSCC enforces the agreement of the instanatiation
policy for the on-disk chaincode when the chaincode container info is
retrieved, which occurs at every invocation, even when retreiving
chaincode container info is unnecessary.  This side-effect style
security check is inhibiting refactoring and needs to be made explicit.

Change-Id: I682a94925106e3802ff5a35f09e8eba9663b0099
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>