change old block's pvtData commit order
Currently, the commit of pvtData of old block execute the
following steps:
 (1) commit the pvtData of old blocks to the pvtdata store and
     a marker denoting the set of pvtData being added.
 (2) commit the data to stateDB.
 (3) remove the marker from the pvtdata store. If the peer fails
     after storing the data in pvtdatastore, the marker helps in
     synching the stateDB with the pvtdataStore.

This flow is mainly used to support failure/recovery.

This can simplify this flow as follows (by removing the marker):
  (1) store the pvtData of old blocks to the stateDB.
  (2) store the pvtData of old blocks to the pvtDataStore.
If the peer fails after storing the pvtData in the stateDB, we don't
need to do any special operation to recover the states. The
reconciler would fetch the data as the pvtData is missing in
the pvtdata Store. While committing the pvtData to the stateDB,
at the worst case, we overwrite existing data but there is no
consistency issues.

FAB-16104 #done

Note:
During a migration from older version to new peer version,
we would need to check for the marker and do necessary action.
Hence, the recovery code still need to have the logic related
to the marker stored in pvtdataStore.

Signed-off-by: senthil <cendhu@gmail.com>
Change-Id: Id931acda740144f8dce4f6f25b7c1331c0f2de6b