FAB-16158 Pass chaincode name instead of ID
The endorser.go code labors under the false assumption that a user
passes in anything in the chaincode ID other than name.  It then goes
and overrides these values anyway.  There's no point to passing the ID
around when the only piece of info it contains is the string name.
Fixing.

Change-Id: If54546736b7927182420304706734420e2181e14
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>