FAB-16158 Pass in chaincode name to Execute
This CR fixes a long-standing problem with the chaincode code and
simplifies things considerably in the process.

For reasons that are elusive, both the endorser.go logic checked
chaincode instantiation policy, as did handler.go, and both separately
retrieved the chaincode definition, to set version information that was
ignored.  This CR consolidates those two paths, and rids us of a large
chunk of code in the process.

Change-Id: Iea94f67068d1fc3013c5a8363bbb52edd78bc1ea
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>