FAB-16158 Remove ContainerInfo as dep
Because package info was passed as a ChaincodeContainerInfo struct, it
managed to find its way through most of the container code.  However,
this struct is ill suited for the realities of building/launching
containers, and mostly duplicates the package metadata.  This CR
switches to using the package metadata and removes the container info as
a dependency.

Change-Id: I983c69d4296f2b4a41c712eb276b5eade4205d82
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>