FAB-16158 Remove persistence/intf type
The persistence/intf type was presumably added in order to help
differentiate between the case of 'chaincode ids' which the chaincode
subsystem uses to register running chaincodes, and for 'package ids',
which are the handles by which we retrieve chaincode packages from the
filesystem.  Now, handling of the package retrieval is outside the scope
of chaincode registration, so this artificial distinction is no longer
necessary and this can be replaced with a vanilla string.  Coming soon,
removal of the other half, namely ccintf.CCID.

Change-Id: Iaa0483e783bf542cb52ebe1d36dde5b8bff55c19
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>