FAB-16278 Name chaincode package .tar.gz
Today, we persist the chaincode packages on disk as .bin files.  But, in
reality, these are .tar.gz files, and, we seem to have made a fairly
strong commitment to this being the packaging format at this point, so
there is nothing really to be gained in obfuscating the filetype.

Change-Id: I73f00d47240c24210ca5a3e4509a3ecf403dd4e3
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>