FAB-16280 Make SysCC Ver constant, add CCID func
Historically, the version of all system chaincodes has been set
dynamically to be the build version of the peer.  This doesn't really
provide us any benefit, other than artificially failing endorsement
requests across syscc versions (such as the new _lifecycle).  Further,
we are creating a CCID for these system chaincode in an ad-hoc fashion
in multiple places.  This CR consolidates the logic and fixes the
version.

Change-Id: I32af5ade97faada12ef138e26ded5378db02fa1d
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>