FAB-16158 Move CCContext to be test only artifact
No one outside of the core/chaincode package referred to the CCContext
type, and this type was only used to pass 3 fields as a return from one
function to a parameter of another.

This CR gets rid of One of these fields (InitRequired) through a slight
refactor, and then passes the other two (name and version) as explicit
parameters.

Despite the parameter lists becoming a little bit long, this still seems
to be the cleaner approach.  If we would like to consolidate parameters,
I would consider a larger unifying structure, though coming up with one
that makes sense was not obvious during this reading.

Change-Id: I5e8dd26fbec7864fdbe4a8e4cd263327778ebcc2
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>