FAB-16166 Remove old metadata filename
This is yet another CR in the back-and-forth series between the peer and
the SDKs to clean up some of the chaincode packaging.  Particularly,
this CR removes the fallback of the old metadata file name now that:

https://gerrit.hyperledger.org/r/c/fabric-sdk-java/+/32770
https://gerrit.hyperledger.org/r/c/fabric-sdk-node/+/32769

have been merged.  Additionally, it removes the annotations from the
metadata struct, which causes the unmarshaling to treat the keys in a
case insensitive way.  This will allow a transition period for the SDKs
to continue working with upper-case metadata keys, while we transition
them to lower case.  Once the transition is complete, we will add back
correct annotations.

Change-Id: Ifbc4f13bbcfb2f4e5454bb7208113164481726f5
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>