Hygiene: Kill dead code in endorser.go
As the chaincode package has been cleaned up, it's becoming more and
more obvious that many code paths and checks are dead or otherwise
no-ops.  For instance, in this CR, the client is expected to send a
ChaincdoeDeploymentSpec, with a nil code package.  We then validate that
nil code package against the packaging registry, sanitize the code
package to ensure that it is nil and only the input is preserved, then
reference only the input in the chaincode package.  All of this code
does nothing and can be removed.

Change-Id: I63bd76225429b5f977c6e0532ddb7d9b47da51f5
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>