FAB-16353 Introduce unpacked proposal struct
Currently, the endorser calls validation code, which unpacks a proposal,
and returns a subset of the bits it unpacked as multiple return values.
Then, the endorser re-unpacks some of those bits, duplicating error
checking, etc.  This CR reworks the validation flow to do all the
unpacking first, then do the required validations.

This CR also adds tests to cover the unpacking path.  More rework to
come.

Change-Id: I4126d8073fa4681540c4561eb95e408fb7bac67d
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>