FAB-16353 Move proposal parsing out of eplugin
Presently, the endorser plugin contract does some static parsing and
error handling for the proposal which is duplicative in some ways of
what the endorser.go validation does.  Because this logic is not per
plugin, there is no reason it needs to be tied with the plugin provider.
So, this CR moves it back into endorser.go so that the common logic can
be factored out.

Change-Id: Ic9451b44eba5c44a3e8ba0bcfb9d409454b68b00
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>