FAB-16353 Compute proposal hash as part of parsing
There are multiple paths where we might need the proposal hash, and it
feels more like a property of a well formed proposal than something that
should be possible to error on later in the flow.  This will hopefully
allow us to treat failure responses and success responses more
uniformly down the line.

Change-Id: Iaf56d62ab483fe542266aa845b3da032b4ebcb1d
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>