FAB-16115 Remove FabToken
FabToken was a good experiment for testing a
potential way of implementing token functionality.

However, the current implementation of the feature
is not going to be included moving forward.

This change removes the current FabToken
implementation from the code base.

Note that the proto definitions were changed
to included reserved fields.  This is because
the token code was also in the v1.4.x protos.

FAB-16115 #done

Change-Id: I519a34a961239de2be644738661008f281682d80
Signed-off-by: Gari Singh <gari.r.singh@gmail.com>