FAB-16353 Explicitly pass the MSP to validate
This CR has a lot of fallout, notably, it breaks all of the existing
unit tests for endorser.go, as it relies on the global singleton method
of mocking via util functions.  Although the diff is quite large,
there's really very little production code changed, it's almost entirely
rewriting the test suite in ginkgo.

Change-Id: I9160f43f7a4d1098a5c9f42de58626ee934dd7e4
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>