FAB-16353 Move off all protoutil.Get methods
The endorser message validation still relied on some of the incredibly
opaquely named 'Get*' functions, some of which unmarshaled, some of
which marshaled, some of which performed additional validation.

This CR transitions the endorser msg validation code to use only the
protoutil.Unmarshal utilities, which pre-decorate the error message with
the proto type, so additionally this CR removes the wrapping message.

Change-Id: If8581dec95b694111f7349a0fcaa83163eb0fc95
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>