FAB-13528 Fix potential container file-handle leak
When there are is no external builder pointer set (which, in production
code today is always set), then an invocation of 'Build' would lead to
leak a filehandle.  In general, for a healthy chaincode, 'Build', is
only invoked once per process, so likely this would have never caused
real problems.  So this is barely a bug, that isn't even currently
exhibited, but it should still be fixed.

Change-Id: Ibc895c50637d0e122717cf6648046b4398201fb8
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>