Broken Build : Spaces in doc filename
A recently checked in doc change included a powerpoint file which has
spaces in its name.  This seems to be confusing the make implicit target
parsing for some reason.  This seems like a bug in our Make structure,
but for now, the easiest thing is simply to rename the file.

Change-Id: I6e01cb86a0b4723d509b1c10f7a0c3483eac167f
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>