FAB-7559 Add orderer connection source
Today, the deliver service gets its connection info from two disparate
and separately updated sources.  Firstly, the gossip eventer gets
updates as to the orderer addresses, secondly, the comm producer code
gets updates for orderer TLS CAs.  The deliver service then combines
these two via a factory to produce connections in the block provider
broadcast client to opaquely connect to the orderer.

Although this deserves a more broad rework, the first obvious
improvement is to have only a single update path for information related
to the ordering connections.  This CR provides an initial implementation
(that is unwired) to track a set of orderer endpoints, update them, and
invalidate the old ones on update.  This already addresses some corner
cases which were not handled in the earlier code.  For instance, not
only when orderer addresses change, but also when TLS CAs change, the
deliver broadcast client should disconnect and reconnect.

In the coming CRs, this new connection source will be enhanced with the
necessary function to support the deliver broadcast client, and will
ultimately be wired through, allowing the other bifurcated
implementation to be removed.

Change-Id: I8b793a682873fb2d84c8de41cc6db29bf8599af4
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>