[FAB-16504] pause and resume a channel
Add the ability to pause/resume channels when a peer is stopped.
When a channel is paused on a peer, the peer will not start
resources for the channel and will not receive new blocks
or process requests for the channel. When a channel is resumed,
the peer will receive blocks and process channel requests.

Change-Id: I15e0633dfb3311052467d7d798ac3807fe14f603
Signed-off-by: Wenjian Qiao <wenjianq@gmail.com>