FAB-16353 Enforce check on nil code-package
The old endorser code nil-ed the code package of the chaincode
deployment spec before passing it on to the legacy chaincode init.
Although none of the SDKs or CLI should be setting the code package,
this CR adds a check to enforce that any other clients fo the old
lifecycle do not accidentally supply a code package.

Change-Id: I9b4a127e0efa6f659dc68d1a25847ecd6e0eabd0
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>