FAB-7559 Convert comm host override to opt
The deliver service work requires the ability to explicitly set a TLS
cert pool on the connection.  Rather than extend the NewConnection
function with yet another TLS option, it's better to follow the gRPC
options pattern of supplying optional opt functions.  This CR does the
conversion for the existing hostname override to prepare to add the
certpool override.

Change-Id: I083a40b89128c12abbacbcb43ae00700208714c7
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>