FAB-13528 Build on install, return err to user
This CR adds to the new lifecycle path, the new semantic that chaincode
packages are built during install, and if an error occurs during the
build, that error is returned to the user.

The quality of the error is still likely very suspect, but it is a step
in the right direction.

Change-Id: I9463967f40928662645c828f2a5fea86bc5453aa
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>