[FAB-16564] policy evaluator takes id slice as arg
As part of the policy rework, we ensure that the internal evaluator function
for policies takes a slice of identities as argument (as opposed to a slice
of structs containing message, signature and serialised identity).
Deserialisation of the identity, validation and verification of the
signature is performed outside of the internal evaluator function, paving
the way for decoupling the two tasks further.

Change-Id: Ie6ec18f433bbc1acf60b3f4f28c6fa5ed2547203
Signed-off-by: Alessandro Sorniotti <ale.linux@sopit.net>