FAB-16599 Add optional release to externalbuilders
Per the design, the external builders are to be given the opportunity to
supply a 'release' binary, which will be invoked after the build, and
give the chaincode an opportunity to supply assets such as connection
information for a remote chaincode, or chaincode db indicates which must
be installed into the statedb for the chaincode to function properly.

This CR adds rudimentary support for release.

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>
Change-Id: I743174c825eae7fd62eaca03bb2905ddfa7cdc1b