FAB-16109 Make run synchronous in externalbuilders
Today, 'run' is expected to fork off the chaincode and disown it, so
that a status of successful launch can be returned to the peer.
However, it's much simpler if the process is simply launched in a
blocking manner, so that the peer may monitor for process exit to detect
failure and allow it to start.  This CR adopts those latter semantics.

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>
Change-Id: I462fab6755522bae629e3f07bf2addd8c9e16876