FAB-16599 Re-use ccid sanitizing for persistence
Because we want to support legacy chaincodes, and legacy chaincodes
necessarily have a ':' in their CCID, using this to persist the
chaincode data is problematic in test artifacts which must be checked
out on Windows.  This CR sanitizes the CCID of generally harmful
characters.  For the old lifecycle, there is a mild risk of collision,
but with the new lifecycle because of the incorporated hashes, there
should be no collision risk.

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>
Change-Id: Ib0bdbff6a64b55adee98fe8ed4d48d80959c7d09