[FAB-16810] comm test changes to support go 1.13
TLS 1.3 is enabled by default in go 1.13. When TLS 1.3 is used, clients
will not receive handshake errors due to invalid or missing certificates
until they attempt to read data from the transport. This impacts when
and how errors are presented to gRPC clients.

This change sets the max TLS version for a bad-certificate test case to
TLS 1.2 to preserve the existing assertions.

Change-Id: Ie8cd57b6c4c92a761ba5ba11d0008f14e8179e81
Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>