Remove invalid test case
The UT being removed assumes ram ledger, whose capacity can be
manually set. However, that's not the case for file ledger. In
Fabric today, we do not permit ledger pruning.

Furthermore, the second sub-test in it does not actually depend
on invalid config metadata, as it claims to. It passes for exactly
the same reason as the first sub-test: old block being pruned.

FAB-16898 #done

Change-Id: Ie4a86d17c27c28dd7db580f3842ab23006f76861
Signed-off-by: Jay Guo <guojiannan1101@gmail.com>