Fix rebuild of dbs
This CR fixes the bug that prevents rebuilding of dbs.
The bug was introduced with the changes made in deployed-chaincode-info-
provider for new lifecycle. The new deployed-chaincode-info-provider
computes the implicit collections from in-memory channel config and the
channel config is loaded in memory after ledger has started.

This CR removes the need of computing implicit collection from the
function as it was not essential to include implicit collections there
at the first place and was added just for the sake of completeness of
info.

This CR also adds a high level test to make sure that the rebuild of db
does not escape from future changes.

FAB-16923 #done

Change-Id: Ice47363ac5f97f97c84ed24a75ad2942ed293446
Signed-off-by: manish <manish.sethi@gmail.com>