FAB-17001 Delete chaincode package on build fail
The new lifecycle has no existing contract to honor with respect to
installing chaincodes which are not buildable at install time, so it is
odd that we adhere to that contract from the previous LSCC case.

This CR changes the new lifecycle to delete the chaincode from the
filesystem if the build is unsuccessful, under the assumption that
future attempts to build this chaincode will similarly fail, and the
package will simply be taking up space on the filesystem.

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>
Change-Id: I7d12a767dbaf6c9d0b497139c4a99793b4d29a16