FAB-15754 Deprecate solo orderer
This CR removes many Solo references from the docs, it notably does not
touch the chaincode tutorial, or the first network tutorial, as these
are in flux.

It also adds a warning which is emitted when starting a channel with the
Solo consensus type, and removes the default 'SampleInsecureSolo'
profile from configtxgen.

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>
Change-Id: Ide57e8ca4109e7d31d7f363dc088205a6b5683b5