[FAB-17095] MCS not to print identities as bytes
The messageCryptoService prints identities as decimal bytes,
and this doesn't help much with debugging because no sane person
would want to convert a decimal byte array to a PEM manually
and then pass it through openssl to figure out what is the peer
in question.

This change set makes the logs use the String() method
of the PeerIdentityType.

Change-Id: I71706d6d941de3a9fe7d6efbadf28949142232e2
Signed-off-by: yacovm <yacovm@il.ibm.com>